pyproject.toml
[tool.poetry]
name = "api-log-text-analyzer"
version = "0.1.0"
description = "Text/log analyzer that extracts API calls, matches them with API specs, validates requests, and generates curl commands"
authors = ["geonhos <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.15"
# LLM
anthropic = "^0.26.0"
# Vector DB
chromadb = "^0.4.0"
# Embeddings
openai = "^1.0.0"
# Data validation
pydantic = "^2.0"
# CLI
click = "^8.1"
# HTTP client
httpx = "^0.25"
# API Spec parsing
openapi-spec-validator = "^0.7.0"
openapi-pydantic = "^0.4.0"
# Schema validation
jsonschema = "^4.20"
# YAML support
pyyaml = "^6.0"
# Terminal UI
rich = "^13.7"
# Progress bar
tqdm = "^4.66"
# Caching
diskcache = "^5.6"
# Environment variables
python-dotenv = "^1.0"
# Clipboard
pyperclip = "^1.8"
# File watching
watchdog = "^3.0"

[tool.poetry.group.dev.dependencies]
# Code formatting
black = "^23.12"
# Import sorting
isort = "^5.13"
# Linting
flake8 = "^7.0"
ruff = "^0.1.9"
# Type checking
mypy = "^1.8"
# Testing
pytest = "^7.4"
pytest-cov = "^4.1"
pytest-asyncio = "^0.23"
# Code complexity
radon = "^6.0"
# Memory profiling
memory-profiler = "^0.61"

[tool.poetry.group.langchain]
optional = true

[tool.poetry.group.langchain.dependencies]
langchain = "^0.1.0"
langchain-anthropic = "^0.1.0"
langchain-community = "^0.0.20"

# [tool.poetry.group.llamaindex]
# optional = true
#
# [tool.poetry.group.llamaindex.dependencies]
# llama-index = {version = "^0.9.0", python = ">=3.9,<3.12"}
# llama-index-llms-anthropic = {version = "^0.1.0", python = ">=3.9,<3.12"}
# llama-index-vector-stores-chroma = {version = "^0.1.0", python = ">=3.9,<3.12"}

[tool.poetry.scripts]
analyzer = "src.legacy.cli.main:cli"
analyzer-langchain = "src.langchain_version.cli.main:cli"
analyzer-llamaindex = "src.llamaindex_version.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.eggs
  | \.pytest_cache
  | \.mypy_cache
  | __pycache__
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# mypy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "anthropic.*",
    "openai.*",
]
ignore_missing_imports = true

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
